requirements.txt
streamlit
stable-baselines3
gymnasium
ta
alpaca-py
matplotlib
scikit-learn
pandas
numpy